+++ grammar/it/src/rules_amount.rs
@@ -147,7 +147,7 @@ pub fn rules_finance(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
              money_unit!(),
              |a, _, b| {
                  Ok(AmountOfMoneyValue {
-                     value: a.value().value as f32,
+                     value: a.value().value as f64,
                      precision: Exact,
                      unit: b.value().unit,
                      ..AmountOfMoneyValue::default()
